fastapi-app/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = ""
authors = [
    {name = "iamnurih",email = "emailnuri@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "asyncmy (>=0.2.10,<0.3.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "fastapi_app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "ruff (>=0.14.0,<0.15.0)",
    "coverage (>=7.10.7,<8.0.0)",
    "aerich (>=0.9.2,<0.10.0)"
]


[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true